+++ src/app/locales/en/components/space-status-icons.js
@@ -1,8 +1,13 @@
 export default {
-  dataImport: 'Data import',
-  dataUpdate: 'Data update',
-  inProgress: 'in progress...',
-  done: 'done!',
-  enabled: 'enabled',
-  working: 'working',
+  importStatus: {
+    manualMode: 'Manual import <strong>enabled</strong>',
+    autoMode: 'Auto import <strong>enabled</strong>',
+    enqueued: 'Auto import scan <strong>enqueued</strong>',
+    running: 'Auto import scan is <strong>running</strong>',
+    aborting: 'Auto import scan is <strong>aborting</strong>',
+    completed: 'Auto import scan <strong>completed</strong>',
+    failed: 'Auto import scan <strong>failed</strong>',
+    aborted: 'Auto import scan <strong>aborted</strong>',
+    nextScanPart: ', next scan: <strong>{{time}}</strong>',
+  },
 };