+++ CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
 ### Latest changes
 
 * VFS-6784 Better UX of triggering storage edition
+* VFS-6801 Disabled ceasing space support feature in favour of removing space in GUI
+* VFS-6330 Redesigned storage import (formerly space synchronization) configuration
 * VFS-6732 Using new monospace font
 * VFS-6652 Enabled assets fingerprinting for more resources
 * VFS-6697 Auto enable and lock skip storage detection in storage form